refactor: JWT 로직을 auth 패키지로 분리하고 Config 통해 Bean 등록
