+++ src/main/java/roomescape/RoomescapeApplication.java
@@ -2,11 +2,10 @@ package roomescape;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import roomescape.auth.JwtProperties;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-@EnableConfigurationProperties(JwtProperties.class)
+@ComponentScan(basePackages = {"roomescape","jwt"})
 public class RoomescapeApplication {
     public static void main(String[] args) {
         SpringApplication.run(RoomescapeApplication.class, args);
